chore(queue): remove queue from App global
